refactor(event-sourcing): custom evaluation sync is a subscriber, and the dead Customer.io reactors are gone
customEvaluationSync moves from withReactor to a fold-bound withSubscriber
on traceSummary, keeping its registration name so jobs staged before a
deploy dispatch into the new registration, and keeping the reactor-era
delay/ttl/dedup semantics byte-for-byte. The three Customer.io reactors
(trace, evaluation, simulation) were implemented but never registered —
the counting strategy they waited on never landed — so they are deleted
rather than migrated.
